Update Bert-VITS2 fp16

+++ bert_vits2/text/chinese_bert_extra.py
@@ -10,7 +10,7 @@ def get_bert_feature(text, word2ph, tokenizer, model, device=config.system.devic
         for i in inputs:
             inputs[i] = inputs[i].to(device)
         res = model(**inputs, output_hidden_states=True)
-        res = torch.nn.functional.normalize(torch.cat(res["hidden_states"][-3:-2], -1)[0], dim=0).cpu()
+        res = torch.nn.functional.normalize(torch.cat(res["hidden_states"][-3:-2], -1)[0], dim=0).float().cpu()
         if style_text:
             style_inputs = tokenizer(style_text, return_tensors="pt")
             for i in style_inputs:
@@ -18,7 +18,7 @@ def get_bert_feature(text, word2ph, tokenizer, model, device=config.system.devic
             style_res = model(**style_inputs, output_hidden_states=True)
             style_res = torch.nn.functional.normalize(
                 torch.cat(style_res["hidden_states"][-3:-2], -1)[0], dim=0
-            ).cpu()
+            ).float().cpu()
             style_res_mean = style_res.mean(0)
     assert len(word2ph) == len(text) + 2
     word2phone = word2ph
