+++ bert_vits2/text/japanese_bert_v200.py
@@ -26,7 +26,7 @@ def get_bert_feature_with_token(tokens, word2ph, tokenizer, model, device=config
         # for i in inputs:
         #     inputs[i] = inputs[i].to(device)
         res = model(**inputs, output_hidden_states=True)
-        res = torch.cat(res["hidden_states"][-3:-2], -1)[0].cpu()
+        res = torch.cat(res["hidden_states"][-3:-2], -1)[0].float().cpu()
     assert inputs["input_ids"].shape[-1] == len(word2ph)
     word2phone = word2ph
     phone_level_feature = []
